requirements.txt
asgiref                       3.5.2
certifi                       2022.9.14
cffi                          1.15.1
charset-normalizer            2.1.1
cryptography                  38.0.1
defusedxml                    0.7.1
dj-rest-auth                  2.2.5
Django                        4.1.1
django-allauth                0.51.0
django-environ                0.9.0
django-rest-auth              0.9.5
djangorestframework           3.13.1
djangorestframework-simplejwt 5.2.0
idna                          3.4
oauthlib                      3.2.1
pip                           21.1.2
pycparser                     2.21
PyJWT                         2.4.0
python3-openid                3.2.0
pytz                          2022.2.1
requests                      2.28.1
requests-oauthlib             1.3.1
setuptools                    57.0.0
six                           1.16.0
sqlparse                      0.4.2
tzdata                        2022.2
urllib3                       1.26.12
wheel                         0.36.2
